Judge-lora/requirements.txt
torch>=2.3.0
transformers>=4.46.0
peft>=0.13.0
bitsandbytes>=0.44.0
accelerate>=1.0.0
datasets>=3.0.0
trl>=0.11.0
sentencepiece
protobuf
pyyaml
flash-attn>=2.6.0; platform_system == "Linux"
